refactor(dongtai-common): remove RemoteConfigUtils dependency in SecondFallbackReportBody;
